requirements.txt
bleach==2.0.0
decorator==4.1.1
entrypoints==0.2.3
html5lib==0.999999999
ipykernel==4.6.1
ipython==5.5.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
jedi==0.10.2
Jinja2==2.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.1.0
jupyter-core==4.3.0
MarkupSafe==1.0
mistune==0.7.4
nbconvert==5.2.1
nbformat==4.3.0
notebook==5.0.0
numpy==1.13.1
olefile==0.44
pandocfilters==1.4.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.2.1
prompt-toolkit==1.0.14
ptyprocess==0.5.2
Pygments==2.2.0
python-dateutil==2.6.1
PyYAML==3.12
pyzmq==16.0.2
qtconsole==4.3.0
simplegeneric==0.8.1
six==1.10.0
terminado==0.6
testpath==0.3.1
torch==0.3.0.post4
torchvision==0.1.8
tornado==4.5.1
tqdm==4.14.0
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==2.0.0
